4th week 7th tack done. Im going sleep

+++ 4thWeek/4thWeek_6z_vozviedieniie-v-stiepien.py
@@ -7,8 +7,8 @@
 def power(a, n):
     if n == 0:
         return 1
-    return a*power(a, n-1)
+    return a * power(a, n-1)
 
 
-a, n = int(input()), int(input())
+a, n = float(input()), float(input())
 print(power(a, n))
